Refactor React components into their own files in src/components/
