+++ src/components/App.css
@@ -1,4 +1,5 @@
-.App {}
+.App {
+}
 
 .Page-header {
   position: relative;
