+++ src/components/Page/PageIcon.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function PageIcon(props) {
+  return (
+    <div className="Page-header-icon">
+      <span className="Page-header-icon">{props.emoji}</span>
+    </div>
+  );
+}
+
+export default PageIcon;
